requirements.txt
Flask
python-dotenv
google-generativeai
requests
pymongo
flask-cors
certifi